xoxo

+++ src/lib/statistics/histograms/abstract_histogram.hpp
@@ -20,7 +20,7 @@ constexpr BinID INVALID_BIN_ID{std::numeric_limits<BinID>::max()};
 /**
  * Used for HistogramBin heights and distinct counts
  */
-using HistogramCountType = size_t;
+using HistogramCountType = float;
 
 template <typename T>
 struct HistogramBin {
